Add conditional block support
This patch adds the `!if` directive to support conditional blocks.

A conditional block is functionally equivalent to a repetition block
when casting the expression to `bool`:

    !if {meow + mix > 32}
      # ...
    !end

    !repeat {meow + mix > 32}
      # ...
    !end

In other words, a conditional block is a repetition of its contained
items zero or one time.

The expression of a fixed-length number, an LEB128 integer, a variable
assignment, and a repetition may now evaluate to a `bool` value: Normand
converts it to `int` (0 or 1) automatically.

Change-Id: I3b686a3196d1ea8daa5741e78320c031b40abc00
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>