Add macro support
This patch makes Normand support macro definitions and macro expansions.

The macro system doesn't support recursion, and to expand a given macro
by name, it needs to be already defined.

A macro definition may only exist at the root level, not within a group
(including not within a repetition, condition, or other macro definition
which all have implicit groups). The form is:

    !macro name(a, b, c)
        ...
    !end

Parameters are optional, but you always need the parentheses after the
macro name.

Parameter names must be unique for a given macro definition.

A parameter becomes a variable for the items of the macro group.

A macro expansion has the form:

    m:name(1, 2, 3)

Again, you always need the parentheses.

To expand a macro, its definition needs to exist completely before (no
recursion).

A parameter value is one of:

* A constant integer.
* A variable/label name.
* An expression between `{` and `}`.

The initial state of a macro expansion is:

Current offset:
    Copied

Current byte order:
    Copied

Variables:
    Parameter values (using parameter names)

Labels:
    None

The final state, after handling a macro expansion, is:

Current offset:
    Incremented by the size of the generated macro expansion data,
    _not_ its final current offset.

Current byte order:
    The one before the expansion.

Variables:
    The ones before the expansion.

Labels:
    The ones before the expansion.

Change-Id: I493c89009d7f7ce61b22f04e70881b253429c3b8
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>