Make it possible to specify more that one byte with `%`
This patch makes it possible to specify more than one consecutive bytes
in binary by using two or more `%` prefixes. The number of `%` indicates
the number of subsequent bytes.

This makes it possible to put a 32-bit binary constant, for example,
without any delimiter. Instead of:

    %11011101 %11110010 %01001101 %11101101

you may write:

    %%%%11011101111100100100110111101101

I don't have a choice to use something (number of `%` here) to indicate
the number of bytes after that because this won't work:

    %11011101111100100100110111101101

This last example is equivalent to:

    %11011101 11 11 00 10 01 00 11 01 11 10 11 01

which really is 13 bytes.

This patch also fixes `normand.py` for Python 3.4.

Change-Id: I4945c9fc5925ab7a32c9015c41b6593db893cdd3
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>