Add transformation block support
A transformation block represents the transformed bytes of one or more
items, for example:

    !transform base64
      "hello how are you"
      {ICITTE:8} * 20
    !end

    61 47 56 73 62 47 38 67  61 47 39 33 49 47 46 79  ┆ aGVsbG8gaG93IGFy
    5a 53 42 35 62 33 55 52  45 68 4d 55 46 52 59 58  ┆ ZSB5b3UREhMUFRYX
    47 42 6b 61 47 78 77 64  48 68 38 67 49 53 49 6a  ┆ GBkaGxwdHh8gISIj
    4a 41 3d 3d                                       ┆ JA==

It's effectively calling a function with the bytes of the block items.

As of this version, Normand only features a specific set of
transformation functions. In the future the user should be able to
define its own transformations (named functions passed to
normand.parse() I guess).

The gzip and bzip2 formats include a timestamp, so the `tests/*.nt`
approach doesn't work here: adding `tests/test_trans_gz_bz2.py` which
uses normand.parse() to compress some data D, and then decompresses the
result using the correct module and compares to D.

Change-Id: I31eb296fa9ba726ee9695cc1cf049abd7cabaf52
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Reviewed-on: https://review.lttng.org/c/normand/+/11030
Tested-by: jenkins <jenkins@lttng.org>