Implement REUSE with SPDX identifiers
The SPDX identifiers [1] are a legally binding shorthand, which can be
used instead of the full boiler plate text. The REUSE specification [2]
defines a standardized method for declaring copyright and licensing for
software projects.

The objective is to facilitate copyright and licensing audits and
compliance for users of our software. We are currently in the process of
converting our projects to REUSE and this will facilitate the vendoring
of normand in those projects.

For files that lacked licensing information, I used the following
guidelines. For code and build system use 'MIT', for documentation
'CC-BY-SA-4.0' and for data files 'CC0-1.0'. This can be altered to your
liking.

Changes to the original patch by Philippe:

* Keep the original copyright notice and MIT license text in
  `normand/normand.py` to keep this module file portable and independent
  from the project.

* Remove newlines between SPDX lines.

[1] https://spdx.org/ids-how
[2] https://reuse.software/tutorial/

Change-Id: I58861cb873585dc773f3993338766c520bf92747
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Reviewed-on: https://review.lttng.org/c/normand/+/11135
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>